FIX: some VS compiler warnings
git-svn-id: https://www.nitrc.org/svn/cmtk/trunk@751 42a5c34f-2066-0410-bec5-ba365beb4995

+++ core/apps/randomwarp.cxx
@@ -65,7 +65,7 @@ main( const int argc, const char *argv[] )
   cmtk::SplineWarpXform *warp = new cmtk::SplineWarpXform( volume, gridSpacing );
 
   // seed RNG with (supposedly) random system time
-  cmtk::MathUtil::NormalRandom( sigma, time( NULL ) );
+  cmtk::MathUtil::NormalRandom( sigma, static_cast<unsigned int>( time( NULL ) ) );
 
   for ( int outFileIdx = 4; outFileIdx < argc; ++outFileIdx ) 
     {
